update: bugfixes... and docs

+++ devtools/tasks/buildFunctions.js
@@ -13,29 +13,29 @@ const fs = require("fs-extra"),
     buildTempPath = path.resolve(distPath, "build/");
 
 /**
- * copy files to the given destination
- * @param {String} sourcePortalPath source of the built portal
- * @param {String} distPortalPath destination folder for the built portal
- * @returns {void}
- */
-function copyPortalFiles (sourcePortalPath, distPortalPath) {
-    fs.copy(sourcePortalPath, distPortalPath).then(() => {
-        console.warn("NOTE: Copied \"" + sourcePortalPath + "\" to \"" + distPortalPath + "\".");
-        replaceStrings(distPortalPath);
-        console.warn("NOTE: Copied \"" + buildTempPath + "\" to \"" + distPortalPath + "\".");
-    }).catch(error => console.error(error));
-}
-
-/**
  * remove files if if they already exist.
- * @param {String} sourcePortalPath source of the built portal
- * @param {String} distPortalPath destination folder for the built portal
+ * @param {Array} allPortalPaths all source paths of all portals to be built
  * @returns {void}
  */
-function removeOldBuiltFiles (sourcePortalPath, distPortalPath) {
+function buildSinglePortal (allPortalPaths) {
+    let sourcePortalPath = [];
+
+    if (allPortalPaths.length === 0) {
+        return;
+    }
+
+    sourcePortalPath = allPortalPaths.pop();
+
+    const portalName = sourcePortalPath.split(path.sep).pop(),
+        distPortalPath = path.resolve(distPath, portalName);
+
     fs.remove(distPortalPath).then(() => {
-        console.warn("NOTE: Deleted directory \"" + distPortalPath + "\".");
-        copyPortalFiles(sourcePortalPath, distPortalPath);
+        // console.warn("NOTE: Deleted directory \"" + distPortalPath + "\".");
+        fs.copy(sourcePortalPath, distPortalPath).then(() => {
+            // console.warn("NOTE: Copied \"" + sourcePortalPath + "\" to \"" + distPortalPath + "\".");
+            replaceStrings(distPortalPath);
+            buildSinglePortal(allPortalPaths);
+        }).catch(error => console.error(error));
     }).catch(function (err) {
         throw new Error("ERROR", err);
     });
@@ -63,19 +63,23 @@ module.exports = function buildWebpack (answers) {
         .map(name => sourcePortalsFolder + "\\" + name)
         .filter(name => fs.lstatSync(name).isDirectory() && !name.endsWith(".git"));
 
-    console.warn("NOTICE: executing command \"" + cliExecCommand + "\"");
+    // console.warn("NOTICE: executing command \"" + cliExecCommand + "\"");
+    console.warn("NOTICE: Building portals. Please wait...");
+
     execute(cliExecCommand).then(function (result) {
+
         console.warn(result.stdout);
+
         prependVersionNumber(path.resolve(buildTempPath, "js/masterportal.js"));
 
         fs.remove(mastercodeVersionPath).then(() => {
-            console.warn("NOTE: Deleted directory \"" + mastercodeVersionPath + "\".");
+            // console.warn("NOTE: Deleted directory \"" + mastercodeVersionPath + "\".");
 
             fs.copy("./img", mastercodeVersionPath + "/img").then(() => {
-                console.warn("NOTE: Copied \"./img\" to \"" + mastercodeVersionPath + "\".");
+                // console.warn("NOTE: Copied \"./img\" to \"" + mastercodeVersionPath + "\".");
 
                 fs.copy(buildTempPath, mastercodeVersionPath).then(() => {
-                    console.warn("NOTE: Copied \"" + buildTempPath + "\" to \"" + mastercodeVersionPath + "\".");
+                    // console.warn("NOTE: Copied \"" + buildTempPath + "\" to \"" + mastercodeVersionPath + "\".");
 
                     fs.remove(buildTempPath).catch(error => console.error(error));
                 }).catch(error => console.error(error));
@@ -84,12 +88,7 @@ module.exports = function buildWebpack (answers) {
             throw new Error("ERROR", err);
         });
 
-        allPortalPaths.forEach(sourcePortalPath => {
-            const portalName = sourcePortalPath.split(path.sep).pop(),
-                distPortalPath = path.resolve(distPath, portalName);
-
-            removeOldBuiltFiles(sourcePortalPath, distPortalPath);
-        });
+        buildSinglePortal(allPortalPaths);
 
     }).catch(function (err) {
         throw new Error("ERROR", err);
